add LED flag
