+++ EasyCon2/Global/Board.cs
@@ -4,15 +4,15 @@
     {
         public readonly string DisplayName;
         public readonly string CoreName;
-        public readonly int Version = 0x45;
+        public readonly int Version = 0x46;
         public readonly int DataSize;
 
         public static Board[] SupportedBoards = new Board[] {
-            new ("Arduino UNO R3", "UNO", 400),
-            new ("Beetle", "Beetle", 900),
-            new ("Leonardo", "Leonardo", 900),
-            new("Teensy 2.0", "Teensy2", 900),
-            new("Teensy 2.0++", "Teensy2pp", 900),
+            new ("Arduino UNO R3", "UNO", 412),
+            new ("Beetle", "Beetle", 924),
+            new ("Leonardo", "Leonardo", 924),
+            new("Teensy 2.0", "Teensy2", 924),
+            new("Teensy 2.0++", "Teensy2pp", 924),
             };
 
         public Board(string displayname, string corename, int datasize)
